Remove "Gladiator" from rotations.
Listen, there's been a lot of dispute whether or not to remove this. The
truth is, it is a terrible playing map, and if released now it would be
taken as a joke map. I cannot stand to have a map that took 5 minutes to
make in rotations, and albeit MCZone it still is connected to the
database and ruins a lot for its design. The layout offers no tactics
and it's just run to the middle and die, run to the middle and die.

I've asked people in-game, and the majority seemed to be fine with its
removal (though there were ones who did not want it gone, as there
always is). If released now, even, it'd still be denied on the spot and
apply to what I said. This can no longer last with it's poor gameplay.

All what was done for it was //cyl, //cyl, //cyl, then a fountain and a
spawn. It doesn't look attractive to players and if a select amount of
regulars are fine with it being gone, I am too.